refactor: extract app locale helpers to :common:android
Moves Context.withAppLocale() and getCurrentLocaleTag() out of
LanguageUtil so non-AnkiDroid modules (broadcast receivers, services,
widgets) can apply the in-app locale to a Context without depending
on :AnkiDroid.
