+++ AnkiDroid/src/main/java/com/ichi2/utils/LanguageUtil.kt
@@ -19,7 +19,6 @@ import android.content.Context
 import android.content.res.Configuration
 import android.content.res.Resources
 import androidx.annotation.StringRes
-import androidx.appcompat.app.AppCompatDelegate
 import androidx.core.os.ConfigurationCompat
 import androidx.fragment.app.Fragment
 import com.ichi2.anki.R
@@ -27,7 +26,6 @@ import com.ichi2.anki.common.android.appContext
 import com.ichi2.anki.common.preferences.sharedPrefs
 import com.ichi2.anki.compat.CompatHelper
 import net.ankiweb.rsdroid.BackendFactory
-import timber.log.Timber
 import java.util.Locale
 
 /**
@@ -364,43 +362,12 @@ object LanguageUtil {
         return createConfigurationContext(configuration).resources.getString(stringRes, *formatArgs)
     }
 
-    /**
-     * Returns a [Context] with resources using the app language.
-     *
-     * Needed for resources accessed outside an Activity (e.g. from a [BroadcastReceiver][android.content.BroadcastReceiver]
-     * or [Service][android.app.Service]):
-     *
-     * On API < 33, [AppCompatDelegate.setApplicationLocales] only applies to Activity contexts, so
-     *  resources resolve in the system locale.
-     *
-     * Returns [this] unchanged (System language) when no in-app language is set.
-     *
-     * This method will not throw.
-     */
-    fun Context.withAppLocale(): Context =
-        try {
-            val tag = getCurrentLocaleTag()
-            if (tag.isEmpty()) return this
-            val configuration = Configuration(resources.configuration)
-            configuration.setLocale(Locale.forLanguageTag(tag))
-            return createConfigurationContext(configuration)
-        } catch (e: Exception) {
-            Timber.w(e, "withAppLocale")
-            return this
-        }
-
     /** @return string defined with [stringRes] on the specified [locale] */
     fun Fragment.getStringByLocale(
         @StringRes stringRes: Int,
         locale: Locale,
         vararg formatArgs: Any,
     ): String = requireContext().getStringByLocale(stringRes, locale, *formatArgs)
-
-    /**
-     * This should always be called after Activity.onCreate()
-     * @return locale language tag of the app configured language
-     */
-    fun getCurrentLocaleTag(): String = AppCompatDelegate.getApplicationLocales().toLanguageTags()
 
     /**
      * Returns the character to use when separating a list; `, ` in English
